refactor(components): Update invoked components
