+++ app/(auth)/forgot-password/_password-confirmation/page.tsx
@@ -1,5 +1,5 @@
-import SubmitButton from '@/components/button/SubmitButton'
-import FieldComponent from '@/components/form/Field'
+import SubmitButton from '@/components/ui/button/SubmitButton'
+import FieldComponent from '@/components/ui/form/Field'
 import { NextPage } from 'next'
 import Link from 'next/link'
