+++ app/(auth)/sign-up/page.tsx
@@ -4,8 +4,8 @@ import { NextPage } from 'next'
 import { useState, FormEvent } from 'react'
 import { useRouter } from 'next/navigation'
 import Link from 'next/link'
-import FieldComponent from '@/components/form/Field'
-import SubmitButton from '@/components/button/SubmitButton'
+import FieldComponent from '@/components/ui/form/Field'
+import SubmitButton from '@/components/ui/button/SubmitButton'
 import api from '@/app/api/axios'
 import ErrorResponse from '@/app/api/error'
